+++ gradle/libs.versions.toml
@@ -18,7 +18,7 @@ lifecycle-runtime-ktx = "2.6.1"
 activity-compose = "1.7.2"
 # https://github.com/google/ksp/releases
 ksp = "1.8.22-1.0.11"
-horologist = "0.4.8"
+horologist = "0.4.11"
 
 # kotlinx
 collections-immutable = "0.3.5"
